Add patient.
